Moved validation exception to common
